fix a weird bug after rebase?
